+++ src/main/scala/com/chinawiserv/wsmp/occupancy/store/mem/package.scala
@@ -15,9 +15,9 @@ import scala.collection.mutable.{ArrayBuffer, ListBuffer, Map}
   */
 package object mem {
 
-  private[occupancy] def load: Map[String, Map[Int, ArrayBuffer[Byte]]] ={
+  private[occupancy] def load: Map[String, Map[Int, ArrayBuffer[Short]]] ={
     println(DateTime.getCurrentDate_YYYYMMDDHHMMSS);
-    val OCCUPANCY_MEM_DATA = Map[Int, ArrayBuffer[Byte]]();
+    val OCCUPANCY_MEM_DATA = Map[Int, ArrayBuffer[Short]]();
     val time =  DateTime.getCurrentDate_YYYYMMDDWithOutSeparator;
     val year = time.take(TIME_YEAR_LENGTH);
     val daytime = time.takeRight(TIME_DAY_LENGTH);
@@ -28,19 +28,19 @@ package object mem {
     records.foreach(record => {
       println(record);
       val station = record.getInteger("station").toInt;
-      val maxLevels = ArrayBuffer[Byte]();
-      maxLevels ++= JavaConversions.asScalaBuffer[Int](record.get("maxLevels", classOf[java.util.List[Int]])).map(_.toByte);
+      val maxLevels = ArrayBuffer[Short]();
+      maxLevels ++= JavaConversions.asScalaBuffer[Int](record.get("maxLevels", classOf[java.util.List[Int]])).map(_.toShort);
       OCCUPANCY_MEM_DATA += (station -> maxLevels);
     });
     println(DateTime.getCurrentDate_YYYYMMDDHHMMSS);
-    Map[String, Map[Int, ArrayBuffer[Byte]]](time -> OCCUPANCY_MEM_DATA);
+    Map[String, Map[Int, ArrayBuffer[Short]]](time -> OCCUPANCY_MEM_DATA);
   }
 
-  private[occupancy] def getOccupancyRate(time: String, thresholdVal: Byte): List[Document] = {
+  private[occupancy] def getOccupancyRate(time: String, thresholdVal: Short): List[Document] = {
     val records = new ListBuffer[Document]();
     if (StringUtils.isNotBlank(time) && time.length == (TIME_YEAR_LENGTH + TIME_DAY_LENGTH)) {
       val daytime = time.takeRight(TIME_DAY_LENGTH);
-      val OCCUPANCY_MEM_DATA = OCCUPANCY_MEM.toMap.getOrElse(time, Map[Int, ArrayBuffer[Byte]]());
+      val OCCUPANCY_MEM_DATA = OCCUPANCY_MEM.toMap.getOrElse(time, Map[Int, ArrayBuffer[Short]]());
       for((station, maxLevels) <- OCCUPANCY_MEM_DATA if maxLevels != null){
         val totalNum = maxLevels.length;
         val occupancyNum = maxLevels.count(_ > thresholdVal).toDouble;
